refactor: move models to the version scope folder
