+++ src/transformers/v2/dataTypes.spec.ts
@@ -1,6 +1,6 @@
 import { expect } from 'chai';
 import { dataTypeResolver } from './dataTypes';
-import { Schema } from '../../models/Schema';
+import { Schema } from './models/Schema';
 import { anchor } from '../../lib/anchor';
 
 const fixture = [
