+++ src/transformers/v2/pathResponses.ts
@@ -1,5 +1,5 @@
 import { OpenAPIV2 } from 'openapi-types';
-import { Schema } from '../../models/Schema';
+import { Schema } from './models/Schema';
 import { dataTypeResolver } from './dataTypes';
 import { Markdown } from '../../lib/markdown';
 import { transformHeaders } from './headers';
